move offset prop back where it belongs
